Updated Enemy Randomzier to Alpha 6
This update corrects the current issue causing lots of support tickets. Namely, the options were all loading disabled for new users. This has been corrected. It also fixes an issue with a zote enemy that was causing sound effects to overload, fixes a few other things, and adds a new module.